+++ regression/reference/classes/pyclassesmodule.cpp
@@ -402,6 +402,21 @@ initclasses(void)
     Py_INCREF(&PY_Singleton_Type);
     PyModule_AddObject(m, "Singleton", (PyObject *)&PY_Singleton_Type);
 
+    // Shape
+    PY_Shape_Type.tp_new   = PyType_GenericNew;
+    PY_Shape_Type.tp_alloc = PyType_GenericAlloc;
+    if (PyType_Ready(&PY_Shape_Type) < 0)
+        return RETVAL;
+    Py_INCREF(&PY_Shape_Type);
+    PyModule_AddObject(m, "Shape", (PyObject *)&PY_Shape_Type);
+
+    // Circle
+    PY_Circle_Type.tp_base = &PY_Shape_Type;
+    if (PyType_Ready(&PY_Circle_Type) < 0)
+        return RETVAL;
+    Py_INCREF(&PY_Circle_Type);
+    PyModule_AddObject(m, "Circle", (PyObject *)&PY_Circle_Type);
+
     {
         // enumeration DIRECTION
         PyObject *tmp_value;