+++ regression/run/classes/main.f
@@ -22,6 +22,7 @@ program tester
   call test_class1_new_by_value
   call test_class1
   call test_singleton
+  call test_subclass
 
   call fruit_summary
   call fruit_finalize
@@ -193,5 +194,26 @@ contains
     call assert_true(obj0 .eq. obj1, "obj0 .eq obj1")
 
   end subroutine test_singleton
+
+  subroutine test_subclass
+    type(Shape) base
+    type(Circle) circle1
+    type(C_PTR) cxxptr
+    integer ivar
+
+    base = Shape()
+    ivar = base%get_ivar()
+    call assert_equals(ivar, 0, "get_ivar")
+
+    circle1 = Circle()
+    ivar = circle1%get_ivar()
+    call assert_equals(ivar, 0, "get_ivar subclass")
+
+    ! Test inherited Shroud generated methods.
+    cxxptr = circle1%get_instance()
+    call assert_true(c_associated(cxxptr), "subclass instance c_associated")
+    call assert_true(circle1%associated(), "subclass instance associated")
+    
+  end subroutine test_subclass
 
 end program tester